fix: move link_to_index_url to indexing_task.py
