+++ tests/test_attachment_stored.py
@@ -11,9 +11,9 @@ from aidial_rag.document_loaders import load_attachment
 from aidial_rag.document_record import DocumentRecord
 from aidial_rag.documents import load_document
 from aidial_rag.errors import DocumentProcessingError, InvalidDocumentError
-from aidial_rag.index_storage import IndexStorageHolder, link_to_index_url
+from aidial_rag.index_storage import IndexStorageHolder
 from aidial_rag.indexing_config import IndexingConfig
-from aidial_rag.indexing_task import IndexingTask
+from aidial_rag.indexing_task import IndexingTask, link_to_index_url
 from aidial_rag.request_context import RequestContext
 from aidial_rag.resources.dial_limited_resources import DialLimitedResources
 from tests.utils.user_limits_mock import user_limits_mock
